+++ controllers/apply/under10k/terms.js
